creates endpoint for /rooms and moves everything from helpers to handlers
