+++ App/app.py
@@ -1,9 +1,8 @@
 from flask import Flask
 from handlers.flask_api import bookit_api
-from helpers.populate_mock_db import insert_empty_time_slots, insert_random_bookings, update_calendar_weeks, create_admin_db
-from helpers.get_workplace_info import get_user_data
+from handlers.populate_mock_db import insert_empty_time_slots, insert_random_bookings, update_calendar_weeks, create_admin_db
+from handlers.get_workplace_info import get_user_data
 from flask_cors import CORS
-from handlers import CONSTANTS as C
 import os
 
 app = Flask(__name__)
