+++ App/handlers/get_workplace_info.py
